+++ src/main/kotlin/no/roedt/ringesentralen/brukere/Brukerendring.kt
@@ -4,4 +4,4 @@ import io.quarkus.runtime.annotations.RegisterForReflection
 import no.roedt.ringesentralen.person.GroupID
 
 @RegisterForReflection
-data class Brukerendring(val personID: Long, val nyGroupId: GroupID, var epostSendt: Boolean)
+data class Brukerendring(val personID: Int, val nyGroupId: GroupID, var epostSendt: Boolean)
